Load swagger configuration from Spring properties. Fixes #66
This method makes use of Maven resource filtering and manual
configuration of properties in Springs application.properties file.
The Swagger configuration is build using Spring `@Value` to get the
configured properties for each.

I left out the TOS, because it was just a placeholder and I also left
the license and URL to license hardcoded.